User system finished.

+++ app/src/main/java/com/example/topnews/utils/RecordHandler.java
@@ -3,35 +3,62 @@ package com.example.topnews.utils;
 
 import android.util.Log;
 
+import com.example.topnews.MainActivity;
+import com.example.topnews.data.model.LoggedInUser;
+
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
+import java.nio.file.Files;
+import java.nio.file.attribute.BasicFileAttributes;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.Vector;
 
 public class RecordHandler {
     Vector<String> records  = new Vector<>();
-    private String home = "/data/user/0/com.example.topnews/cache/";
+    private final static String TAG = "RecordHandler";
+    private final String home = "/data/user/0/com.example.topnews/cache/";
+    private String type;
     private File recordFile;
+    private boolean modify = false;
+    private boolean initial = true;
 
     public RecordHandler(final String type){
-        recordFile = new File(home+type+".txt");
+        modify = true;
+        String thome = home + MainActivity.user.getUserId() + "/";
+        File homedir = new File(thome);
+        if(!homedir.exists()) homedir.mkdirs();
+        this.type = type;
+        recordFile = new File(home + MainActivity.user.getUserId() + "/" +type+".txt");
         load();
+    }
+
+    public void updateUser(){
+        modify = true;
+        records.clear();
+        String thome = home + MainActivity.user.getUserId() + "/";
+        File homedir = new File(thome);
+        if(!homedir.exists()) homedir.mkdirs();
+        recordFile = new File(home + MainActivity.user.getUserId() + "/" +type+".txt");
     }
 
     public void add(final String newsID){
         if(records.contains(newsID)) return;
         // Log.d("Record Add", newsID);
         records.add(newsID);
+        modify = true;
     }
+
     public void remove(final String newsID){
         if(records.contains(newsID)) records.remove(newsID);
     }
     public void save(){
+        if(!modify) return;
         try {
+            Log.d(TAG,"Save"+recordFile.getAbsolutePath());
             recordFile.delete();
             recordFile.createNewFile();
             BufferedWriter writer = new BufferedWriter(new FileWriter(recordFile));
@@ -40,6 +67,8 @@ public class RecordHandler {
                 writer.write( value + "\n");
             }
             writer.close();
+            modify = false;
+            initial = false;
         } catch (Exception e){
             e.printStackTrace();
         }
@@ -48,8 +77,11 @@ public class RecordHandler {
         try {
             if(!recordFile.exists()){
                 // Log.d("Record Error","File Missing.");
+                records.clear();
+                modify = true;
                 return;
             }
+            Log.d(TAG,"Load"+recordFile.getAbsolutePath());
             BufferedReader reader = new BufferedReader(new FileReader(recordFile));
             String line = null;
             while ((line = reader.readLine())!=null){
@@ -57,6 +89,7 @@ public class RecordHandler {
                 // Log.d("Record Load.",records.lastElement());
             }
             reader.close();
+            modify = true;
         }catch (Exception e){
             e.printStackTrace();
         }
@@ -67,5 +100,14 @@ public class RecordHandler {
 
     public int size(){
         return records.size();
+    }
+
+    public long getModifyTime(){
+        if(initial) return -1;
+        try {
+            return recordFile.lastModified();
+        } catch (Exception e){
+            return 0;
+        }
     }
 }
